+++ src/Compute/Compute/ChangeLog.md
@@ -20,6 +20,10 @@
 
 -->
 ## Upcoming Release
+* Added optional parameter `-EdgeZone` to the following cmdlets:
+    - `New-AzVM`
+    - `New-AzVmss`
+    - `New-AzVmssConfig`
 * Updated the `Set-AzVMDiskEncryptionExtension` cmdlet to support ADE extension migration from two pass (version with AAD input parameters) to single pass (version without AAD input parameters).
     - Added a switch parameter `-Migrate` to trigger migration workflow.
     - Added a switch parameter `-MigrationRecovery` to trigger recovery workflow for VMs experiencing failures after migration from two pass ADE.